# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "注册：创建 User + Personal Workspace + Welcome Credits + Session",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/admin/system-update/status": {
      "get": {
        "operationId": "SystemUpdateController_status",
        "summary": "系统更新状态（当前/最新版本、是否可更新）",
        "parameters": [
          {
            "name": "force",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin/system-update"
        ]
      }
    },
...
    "/api/v1/admin/system-update/update": {
      "post": {
        "operationId": "SystemUpdateController_update",
        "summary": "触发后台更新（到最新或指定版本）",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin/system-update"
        ]
      }
    },
    "/api/v1/admin/system-update/rollback": {
      "post": {
        "operationId": "SystemUpdateController_rollback",
        "summary": "触发后台回滚（回退上一个版本，或指定版本）",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackVersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin/system-update"
        ]
      }
    },
    "/api/v1/admin/system-update/operations/{operationId}": {
      "get": {
        "operationId": "SystemUpdateController_operation",
        "summary": "查询一次后台更新/回滚操作的实时进度",
        "parameters": [
          {
            "name": "operationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin/system-update"
        ]
      }
    },
...
  },
  "info": {
    "title": "Enova Creator API",
    "description": "灵动创影 AI Creator SaaS 后端 API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "turnstileToken": {
            "type": "string",
            "description": "Cloudflare Turnstile 验证码 token（启用时必填）"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string"
          },
          "turnstileToken": {
            "type": "string",
            "description": "Cloudflare Turnstile 验证码 token（启用时必填）"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "UpdateVersionDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "example": "1.2.1",
            "description": "目标版本（缺省为最新 stable）"
          }
        }
      },
      "RollbackVersionDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "example": "1.1.2",
            "description": "回退到指定版本；缺省回退到上一个版本"
          }
        }
      },
...
    }
  }
}